+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix service cache when has composed ids
+
 ### 0.21.2 - 2020-08-27
 
  - Fix Repository.already_got_not_found timeout